Replaces general info with no info state
Replaces the `GENERAL_INFORMATION` avalanche risk level with `NO_INFO` to better represent the state when no information is available.

Updates default values in HazardLevels and BulletinViewState to reflect this change, and updates BulletinMapper to return `NO_INFO` when the risk level cannot be determined.

Updates shimmer effect in presentation layer.